moves the sql queries to own module
